Add files via upload
The files in this folder relate to the final assignment of the Principles of Programming module. The assignment asks us to create a working simulcra of an enigma machine in python and use it to crack codes. It then asks for extensions, in which I create a more complex machine that would have resisted the code breaking at blechly park.